# Use parallel Hashmap for ConflictSet -- not much performance difference