--- a/flow/flat_buffers.h
+++ b/flow/flat_buffers.h
@@ -180,6 +180,31 @@ struct vector_like_traits<std::array<T, N>> : std::true_type {
 	}
 };
 
+#include "parallel_hashmap/phmap.h"
+template <class Key, class T, class Compare, class Allocator>
+struct vector_like_traits<phmap::parallel_flat_hash_map<Key, T, Compare, Allocator>> : std::true_type {
+	using Vec = phmap::parallel_flat_hash_map<Key, T, Compare, Allocator>;
+	using value_type = std::pair<Key, T>;
+	using iterator = typename Vec::const_iterator;
+	using insert_iterator = std::insert_iterator<Vec>;
+
+	template <class Context>
+	static size_t num_entries(const Vec& v, Context&) {
+		return v.size();
+	}
+	template <class Context>
+	static void reserve(Vec& v, size_t size, Context&) {}
+
+	template <class Context>
+	static insert_iterator insert(Vec& v, Context&) {
+		return std::inserter(v, v.end());
+	}
+	template <class Context>
+	static iterator begin(const Vec& v, Context&) {
+		return v.begin();
+	}
+};
+
 template <class Key, class T, class Compare, class Allocator>
 struct vector_like_traits<std::map<Key, T, Compare, Allocator>> : std::true_type {
 	using Vec = std::map<Key, T, Compare, Allocator>;